Moved icons to original folder
