+++ package.json
@@ -21,9 +21,11 @@
         "lint:js:fix": "yarn lint:js --fix .",
         "lint:scss:fix": "yarn lint:scss --fix",
         "lint:fix": "yarn lint:js:fix && yarn lint:scss:fix",
-        "compilejs": "./compile.js",
+        "compilejs": "node ./scripts/compile.js",
         "tsc": "tsc",
-        "build-icons": "svgr --template ./src/tokens/icons/svgrTemplate.js --index-template ./src/tokens/icons/indexTemplate.js --config-file ./src/tokens/icons/svgr.config.json ./src/tokens/icons/assets --out-dir ./src/icons",
+        "prebuild-icons": "node ./scripts/rename.js ./src/icons/index.js ./src/icons/tmp.js",
+        "build-icons": "svgr --template ./src/tokens/icons/svgrTemplate.js --config-file ./src/tokens/icons/svgr.config.json ./src/tokens/icons/assets --out-dir ./src/icons",
+        "postbuild-icons": "node ./scripts/rename.js ./src/icons/index.js ./src/icons/assets.js && node ./scripts/rename.js ./src/icons/tmp.js ./src/icons/index.js",
         "index-icons": "create-index ./src/icons/assets -r -x svg && eslint --fix ./src/icons/assets/index.js",
         "build": "yarn build:storybook && yarn build:lib && yarn webpack",
         "prebuild:lib": "rimraf dist/",
