+++ src/components/Accordion/AccordionItem/AccordionItem.stories.jsx
@@ -5,7 +5,7 @@ import markdownNotes from './AccordionItem.md';
 
 import {Accordion} from '~/components/Accordion';
 import {AccordionItem} from './index';
-import Love from '~/icons/assets/Love';
+import Love from '~/icons/Love';
 
 storiesOf('Components|Accordion/AccordionItem', module)
     .addParameters({
