+++ src/components/Chip/Chip.stories.jsx
@@ -6,13 +6,13 @@ import {boolean, select, text, withKnobs} from '@storybook/addon-knobs';
 import markdownNotes from './Chip.md';
 import {Chip, colors} from './index';
 import {capitalize} from '~/__storybook__/utils';
-import CloudIcon from '~/icons/assets/Cloud';
-import DefaultIcon from '~/icons/assets/Apps';
-import DeleteIcon from '~/icons/assets/Delete';
-import FileContentIcon from '~/icons/assets/FileContent';
-import LockIcon from '~/icons/assets/Lock';
-import NoCloudIcon from '~/icons/assets/NoCloud';
-import WarningIcon from '~/icons/assets/Warning';
+import CloudIcon from '~/icons/Cloud';
+import DefaultIcon from '~/icons/Apps';
+import DeleteIcon from '~/icons/Delete';
+import FileContentIcon from '~/icons/FileContent';
+import LockIcon from '~/icons/Lock';
+import NoCloudIcon from '~/icons/NoCloud';
+import WarningIcon from '~/icons/Warning';
 import storyStyles from '~/__storybook__/storybook.scss';
 
 const labelValue = (defaultValue = 'Chip') => text('Label', defaultValue);
