+++ src/components/PrimaryNav/PrimaryNav.stories.jsx
@@ -5,7 +5,7 @@ import markdownNotes from './PrimaryNav.md';
 
 import {PrimaryNav} from './index';
 import {PrimaryNavItemsGroup, PrimaryNavItem, Button, Badge} from '~/components';
-import {Apps, Feather, Setting, Profile, Person, Power, Star, Workflow} from '~/icons/assets';
+import {Apps, Feather, Setting, Profile, Person, Power, Star, Workflow} from '~/icons';
 
 storiesOf('Components|PrimaryNav', module)
     .addParameters({
