+++ src/components/PrimaryNav/PrimaryNavItemsGroup/PrimaryNavItemsGroup.stories.jsx
@@ -5,7 +5,7 @@ import markdownNotes from './PrimaryNavItemsGroup.md';
 
 import {PrimaryNavItemsGroup} from './index';
 import {PrimaryNavItem} from '~/components/PrimaryNav/PrimaryNavItem';
-import Edit from '~/icons/assets/Edit';
+import Edit from '~/icons/Edit';
 
 storiesOf('Components|PrimaryNavItemsGroup', module)
     .addParameters({
