+++ src/icons/assets.js
@@ -1,0 +1,183 @@
+export { default as AbTesting } from './AbTesting'
+export { default as Accessibility } from './Accessibility'
+export { default as Add } from './Add'
+export { default as AddBox } from './AddBox'
+export { default as AddCircle } from './AddCircle'
+export { default as AddCircleOutline } from './AddCircleOutline'
+export { default as AddComment } from './AddComment'
+export { default as AddFile } from './AddFile'
+export { default as AddFolder } from './AddFolder'
+export { default as AddPerson } from './AddPerson'
+export { default as Apps } from './Apps'
+export { default as ArrowDown } from './ArrowDown'
+export { default as ArrowDropDown } from './ArrowDropDown'
+export { default as ArrowDropUp } from './ArrowDropUp'
+export { default as ArrowLeft } from './ArrowLeft'
+export { default as ArrowRight } from './ArrowRight'
+export { default as ArrowUp } from './ArrowUp'
+export { default as Back } from './Back'
+export { default as Bar } from './Bar'
+export { default as BarSquare } from './BarSquare'
+export { default as Book } from './Book'
+export { default as Bookmark } from './Bookmark'
+export { default as Bug } from './Bug'
+export { default as Build } from './Build'
+export { default as Calendar } from './Calendar'
+export { default as Cancel } from './Cancel'
+export { default as Check } from './Check'
+export { default as CheckboxChecked } from './CheckboxChecked'
+export { default as CheckboxMiddle } from './CheckboxMiddle'
+export { default as CheckboxUnchecked } from './CheckboxUnchecked'
+export { default as ChevronBigLeft } from './ChevronBigLeft'
+export { default as ChevronBigRight } from './ChevronBigRight'
+export { default as ChevronDoubleLeft } from './ChevronDoubleLeft'
+export { default as ChevronDoubleRight } from './ChevronDoubleRight'
+export { default as ChevronDown } from './ChevronDown'
+export { default as ChevronLeft } from './ChevronLeft'
+export { default as ChevronRight } from './ChevronRight'
+export { default as ChevronUp } from './ChevronUp'
+export { default as Clock } from './Clock'
+export { default as Close } from './Close'
+export { default as Cloud } from './Cloud'
+export { default as CloudCheck } from './CloudCheck'
+export { default as CloudDownload } from './CloudDownload'
+export { default as CloudUpload } from './CloudUpload'
+export { default as Code } from './Code'
+export { default as Collections } from './Collections'
+export { default as CollectionsBookmark } from './CollectionsBookmark'
+export { default as Comments } from './Comments'
+export { default as Compare } from './Compare'
+export { default as Copy } from './Copy'
+export { default as Crown } from './Crown'
+export { default as Cut } from './Cut'
+export { default as Dashboard } from './Dashboard'
+export { default as DefaultEntry } from './DefaultEntry'
+export { default as Delete } from './Delete'
+export { default as DeletePermanently } from './DeletePermanently'
+export { default as DevTools } from './DevTools'
+export { default as Diagram } from './Diagram'
+export { default as Dns } from './Dns'
+export { default as Download } from './Download'
+export { default as DropLeft } from './DropLeft'
+export { default as DropRight } from './DropRight'
+export { default as Edit } from './Edit'
+export { default as Event } from './Event'
+export { default as Exit } from './Exit'
+export { default as Explore } from './Explore'
+export { default as Feather } from './Feather'
+export { default as File } from './File'
+export { default as FileCompresed } from './FileCompresed'
+export { default as FileContent } from './FileContent'
+export { default as FileImage } from './FileImage'
+export { default as FileZip } from './FileZip'
+export { default as Filter } from './Filter'
+export { default as Folder } from './Folder'
+export { default as FolderOpen } from './FolderOpen'
+export { default as FolderSpecial } from './FolderSpecial'
+export { default as FolderUser } from './FolderUser'
+export { default as Follow } from './Follow'
+export { default as Form } from './Form'
+export { default as Funnel } from './Funnel'
+export { default as GlobalLink } from './GlobalLink'
+export { default as Grain } from './Grain'
+export { default as GraphQl } from './GraphQl'
+export { default as Gravel } from './Gravel'
+export { default as Group } from './Group'
+export { default as GroupWork } from './GroupWork'
+export { default as HandleDrag } from './HandleDrag'
+export { default as HandleMove } from './HandleMove'
+export { default as HandleResize } from './HandleResize'
+export { default as Heal } from './Heal'
+export { default as Help } from './Help'
+export { default as HelpOutline } from './HelpOutline'
+export { default as Hidden } from './Hidden'
+export { default as Home } from './Home'
+export { default as Image } from './Image'
+export { default as Info } from './Info'
+export { default as Information } from './Information'
+export { default as JContent } from './JContent'
+export { default as Jwt } from './Jwt'
+export { default as Label } from './Label'
+export { default as Language } from './Language'
+export { default as Layers } from './Layers'
+export { default as Line } from './Line'
+export { default as Link } from './Link'
+export { default as Loading } from './Loading'
+export { default as LocalSiteManager } from './LocalSiteManager'
+export { default as Lock } from './Lock'
+export { default as Losange } from './Losange'
+export { default as Love } from './Love'
+export { default as ManageSiteWeb } from './ManageSiteWeb'
+export { default as Maximize } from './Maximize'
+export { default as Menu } from './Menu'
+export { default as Minimize } from './Minimize'
+export { default as Module } from './Module'
+export { default as MoreHoriz } from './MoreHoriz'
+export { default as MoreVert } from './MoreVert'
+export { default as NoCloud } from './NoCloud'
+export { default as Not } from './Not'
+export { default as Notices } from './Notices'
+export { default as Notifications } from './Notifications'
+export { default as NotificationsActive } from './NotificationsActive'
+export { default as OpenInBrowser } from './OpenInBrowser'
+export { default as OpenInNew } from './OpenInNew'
+export { default as Palette } from './Palette'
+export { default as Paste } from './Paste'
+export { default as Pdf } from './Pdf'
+export { default as Person } from './Person'
+export { default as PhotoLibrary } from './PhotoLibrary'
+export { default as Pie } from './Pie'
+export { default as Power } from './Power'
+export { default as Print } from './Print'
+export { default as Profile } from './Profile'
+export { default as Publish } from './Publish'
+export { default as Puzzle } from './Puzzle'
+export { default as RadioChecked } from './RadioChecked'
+export { default as RadioUnchecked } from './RadioUnchecked'
+export { default as Reload } from './Reload'
+export { default as RemotePublication } from './RemotePublication'
+export { default as Replay } from './Replay'
+export { default as Reply } from './Reply'
+export { default as Report } from './Report'
+export { default as SdLgenerator } from './SdLgenerator'
+export { default as SdLreport } from './SdLreport'
+export { default as Save } from './Save'
+export { default as School } from './School'
+export { default as Search } from './Search'
+export { default as Section } from './Section'
+export { default as Security } from './Security'
+export { default as Server } from './Server'
+export { default as Setting } from './Setting'
+export { default as Share } from './Share'
+export { default as SiteSettingsPublication } from './SiteSettingsPublication'
+export { default as SiteWeb } from './SiteWeb'
+export { default as Star } from './Star'
+export { default as Studio } from './Studio'
+export { default as Subdirectory } from './Subdirectory'
+export { default as SwapHoriz } from './SwapHoriz'
+export { default as SwapVert } from './SwapVert'
+export { default as Tag } from './Tag'
+export { default as Target } from './Target'
+export { default as Task } from './Task'
+export { default as Text } from './Text'
+export { default as Tune } from './Tune'
+export { default as Undelete } from './Undelete'
+export { default as Undo } from './Undo'
+export { default as UnfoldLess } from './UnfoldLess'
+export { default as UnfoldMore } from './UnfoldMore'
+export { default as Unlink } from './Unlink'
+export { default as Unlock } from './Unlock'
+export { default as Upload } from './Upload'
+export { default as ViewColumn } from './ViewColumn'
+export { default as ViewComfy } from './ViewComfy'
+export { default as ViewGrid } from './ViewGrid'
+export { default as ViewList } from './ViewList'
+export { default as ViewQuilt } from './ViewQuilt'
+export { default as ViewStream } from './ViewStream'
+export { default as Visibility } from './Visibility'
+export { default as Warning } from './Warning'
+export { default as WebPage } from './WebPage'
+export { default as WebProject } from './WebProject'
+export { default as Widgets } from './Widgets'
+export { default as Work } from './Work'
+export { default as Workflow } from './Workflow'
